requirements.txt
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.6
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
confluent-kafka==2.13.0
cryptography==46.0.3
Faker==40.1.0
fastavro==1.12.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
py4j==0.10.9.7
pycparser==2.23
pyspark==3.5.1
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
